# server/services/doctor/swagger.yaml
openapi: 3.0.3
info:
  title: Doctor Service API
  version: 1.0.0
  description: API documentation for doctor endpoints only.
servers:
  - url: http://localhost:4003
    description: Local doctor service

tags:
  - name: Doctor
    description: Doctor management and authentication

paths:
  /api/v1/doctor/register:
    post:
      tags:
        - Doctor
      summary: Register a doctor account
      operationId: registerDoctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDoctorRequest'
      responses:
        '201':
          description: Doctor registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterDoctorSuccessResponse'
        '409':
          description: Doctor already exists with this email or license number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/doctor/login:
    post:
      tags:
        - Doctor
      summary: Login doctor and receive JWT
      operationId: loginDoctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDoctorRequest'
      responses:
        '200':
          description: Doctor login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginDoctorSuccessResponse'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/doctor:
    post:
      tags:
        - Doctor
      summary: Create a doctor record
      operationId: createDoctor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDoctorRequest'
      responses:
        '201':
          description: Doctor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDoctorSuccessResponse'
        '409':
          description: Doctor already exists with this email or license number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Doctor
      summary: List all doctors
      operationId: listDoctors
      responses:
        '200':
          description: Doctor list fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListDoctorsSuccessResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    Doctor:
      type: object
      properties:
        id:
          type: string
          example: clx8d3f9b0000abc123def456
        name:
          type: string
          example: Dr. Sarah Johnson
        email:
          type: string
          format: email
          example: sarah.johnson@hospital.com
        specialization:
          type: string
          example: Cardiology
        role:
          type: string
          enum: [doctor]
          example: doctor
        createdAt:
          type: string
          format: date-time
          example: 2026-04-01T10:15:30.000Z
      required:
        - id
        - name
        - email
        - specialization
        - role
        - createdAt

    CreateDoctorRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          example: Dr. Sarah Johnson
        email:
          type: string
          format: email
          example: sarah.johnson@hospital.com
        password:
          type: string
          minLength: 8
          description: Must include uppercase, lowercase, and number.
          example: SecurePass123
        specialization:
          type: string
          minLength: 2
          example: Cardiology
        licenseNumber:
          type: string
          minLength: 3
          example: LIC-2026-001
        phone:
          type: string
          example: +1-555-123-4567
      required:
        - name
        - email
        - password
        - specialization
        - licenseNumber

    RegisterDoctorRequest:
      allOf:
        - $ref: '#/components/schemas/CreateDoctorRequest'

    LoginDoctorRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: sarah.johnson@hospital.com
        password:
          type: string
          minLength: 1
          example: SecurePass123
      required:
        - email
        - password

    ErrorBody:
      type: object
      properties:
        message:
          type: string
          example: Invalid email or password
        details:
          nullable: true
      required:
        - message

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorBody'
      required:
        - success
        - error

    CreateDoctorSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Doctor created successfully
        data:
          $ref: '#/components/schemas/Doctor'
      required:
        - success
        - message
        - data

    RegisterDoctorSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Doctor registered successfully
        data:
          $ref: '#/components/schemas/AuthDoctorData'
      required:
        - success
        - message
        - data

    LoginDoctorSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Doctor login successfull
        data:
          $ref: '#/components/schemas/AuthDoctorData'
      required:
        - success
        - message
        - data

    AuthDoctorData:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.sample.signature
        doctor:
          $ref: '#/components/schemas/Doctor'
      required:
        - token
        - doctor

    ListDoctorsSuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Doctor'
      required:
        - success
        - data